Refactor ProjectCollaborator model and service to use appropriate data types and enhance user authentication handling

+++ API/src/main/java/com/grabit/app/model/ProjectCollaborator.java
@@ -27,7 +27,7 @@ public class ProjectCollaborator {
     private Integer projectID;
 
     @Column(name = "RoleID", nullable = false)
-    private Integer roleID;
+    private Byte roleID;
 
     @Column(name = "JoinedAt", nullable = false)
     private LocalDateTime joinedAt;
